Enable usage of xdgud as "xdg-user-dir"
The plan for this feature:

When invoked as "xdg-user-dir", the program will exhibit behavior
similar to the xdg-user-dir provided by freedesktop.org. Options will
not be accepted and it will return the proper directory for proper
input.

The difference will be that it returns an error and doesn't produce a
directory when an invalid directory name is provided. The
freedesktop.org version returns the user's home directory for some
inexplicable reason.