Change get_config_home() to be more like getline()
The function still dynamically allocates a string, but does so using an
exponential increase in allocation size instead of allocating the exact
size necessary for the string.
It also allows for an already-allocated string to be passed to the
function instead of always allocating a new one.

The change isn't beneficial in this particular situation and probably
slows it down, but this capability is planned to be used further.